requirements.txt
Django==1.10
psycopg2==2.6.1
uWSGI==2.0.14
sorl-thumbnail==12.3
Pillow==3.3.1
django-admin-tools==0.8.0
djangorestframework==3.4.7
iso3166==0.7
django-solo==1.1.2
django-rest-auth==0.9.0
django-allauth==0.30.0